Feature/user-permissions

+++ src/main/java/com/javaapi/pmanager/domain/applicationservice/ProjectService.java
@@ -4,17 +4,21 @@ import com.javaapi.pmanager.domain.applicationservice.ports.ProjectPublisher;
 import com.javaapi.pmanager.domain.entity.Project;
 import com.javaapi.pmanager.domain.entity.User;
 import com.javaapi.pmanager.domain.events.ProjectCreatedEvent;
+import com.javaapi.pmanager.domain.exception.DeniedUserException;
 import com.javaapi.pmanager.domain.exception.DuplicateProjectException;
 import com.javaapi.pmanager.domain.exception.InvalidProjectStatusExcpetion;
 import com.javaapi.pmanager.domain.exception.ProjectNotFoundException;
 import com.javaapi.pmanager.domain.model.ProjectStatus;
+import com.javaapi.pmanager.domain.model.UserRole;
 import com.javaapi.pmanager.domain.repository.ProjectRepository;
 import com.javaapi.pmanager.infrastructure.dto.SaveProjectDataDTO;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import org.springframework.web.bind.annotation.ExceptionHandler;
 
+import java.awt.*;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
@@ -30,7 +34,10 @@ public class ProjectService {
     private final ProjectPublisher projectPublisher;
 
     @Transactional
-    public Project createProject(SaveProjectDataDTO saveProjectData) {
+    public Project createProject(SaveProjectDataDTO saveProjectData, User currentUser) {
+         if (currentUser.getRoles() != UserRole.ADMIN) {
+            throw new DeniedUserException("ADMIN");
+         }
         if (existsProjectWithName(saveProjectData.getName(), null)) {
             throw new DuplicateProjectException(saveProjectData.getName());
         }
@@ -65,13 +72,19 @@ public class ProjectService {
     }
 
     @Transactional
-    public void deleteProject(String projectId) {
+    public void deleteProject(String projectId, User currentUser) {
+        if (currentUser.getRoles() != UserRole.ADMIN) {
+            throw new DeniedUserException("ADMIN");
+        }
         Project project = loadProject(projectId);
         projectRepository.delete(project);
     }
 
     @Transactional
-    public Project updateProject(String projectId, SaveProjectDataDTO saveProjectData) {
+    public Project updateProject(String projectId, SaveProjectDataDTO saveProjectData, User currentUser) {
+        if (currentUser.getRoles() != UserRole.ADMIN) {
+            throw new DeniedUserException("ADMIN");
+        }
         if (existsProjectWithName(saveProjectData.getName(), projectId)) {
             throw new DuplicateProjectException(saveProjectData.getName());
         }
